(17) refactor: abstract out app.ts from index.ts
